Update return type for uint32_t pointer to long
Changed the return type from int to long for uint32_t pointers to better represent memory addresses. This ensures consistency and prevents potential issues with pointer size mismatches.

+++ buildSrc/src/main/java/com/onemillionworlds/tamarin/gradle/tasks/generators/X10Generator.java
@@ -249,7 +249,7 @@ public class X10Generator extends FileGenerator {
         if (cType.equals("char") && isPointer) {
             return "String";
         } else if (cType.equals("uint32_t") && isPointer) {
-            return "int";
+            return "long"; // Changed from int to long as it's a pointer (memory address)
         } else if (cType.equals("XrSystemId")) {
             return "long";
         } else if (isHandleType(cType)) {
